Moved parser.py to model
