+++ controller/start.py
@@ -5,7 +5,7 @@ sys.path.append("..")
 from controller.title_loop import ten_second_loop
 from controller.argument_parser import ticker_argument_parser
 from view.main_view import MainView
-from model.feed_manager import parse
+from model.parser import *
 
 
 if __name__ == "__main__":
