+++ model/feed_manager.py
@@ -152,21 +152,3 @@ class FeedManager:
         except FeedNotFoundError:
             feed = Feed(feed_name, article_list)
             self.__list_of_feeds.append(feed)
-
-
-def parse(feed_link: str) -> []:
-    # Get the contents of an atom or rss feed using the feedparser library. Return all the relevant
-    #   information as Articles (unsorted).
-
-    feed = feedparser.parse(feed_link)
-    article_list = []
-
-    for entry in feed.entries:
-        title = entry.title
-        link = entry.link
-        datetime = entry.published_parsed  # time.struct_time object parsed within feedparser from string attribute
-        article = Article(title, link, datetime)
-
-        article_list.append(article)
-
-    return article_list
